searchPage 분리하던거 세이브
