+++ FrontEnd/src/components/mainPage/SearchContainer.jsx
@@ -2,9 +2,10 @@ import styled from './SearchContainer.module.css'
 import PropTypes from 'prop-types';
 import { Search } from "lucide-react";
 import { useState } from 'react';
+import { useNavigate } from "react-router-dom";
 
 export default function SearchContainer(props) { 
-
+    const navigate = useNavigate();
     const { searchTerm, setSearchTerm, onSearch} = props;
     const [value, setValue] = useState(searchTerm || ''); 
     
@@ -15,7 +16,8 @@ export default function SearchContainer(props) {
 
     function handleKeyDown(event) {
         if (event.key === "Enter" && value.trim()) {
-            onSearch(); 
+            const keyword = searchTerm;
+            navigate(`/search/${keyword}`);
         }
     }
 
@@ -48,7 +50,6 @@ export default function SearchContainer(props) {
 SearchContainer.propTypes = {
     searchTerm: PropTypes.string.isRequired,   // searchTerm은 string이어야 함
     setSearchTerm: PropTypes.func.isRequired,  // setSearchTerm은 함수여야 함
-    onSearch: PropTypes.func.isRequired,
 };
