+++ FrontEnd/src/pages/MainPage.jsx
@@ -1,23 +1,12 @@
 import styled from './MainPage.module.css';
 import SearchContainer from '../components/mainPage/SearchContainer';
 import MainNews from '../components/mainPage/MainNews';
-import SearchNews from '../components/mainPage/SearchNews';
-import BlankNews from '../components/mainPage/BlankNews';
 
 import { useState } from 'react';
-
-import { getSearch } from '../api/getNewsCardAPI';
 
 export default function MainPage() {
 
     const [searchTerm, setSearchTerm] = useState("");
-    //const [isSearch, setIsSearch] = useState(false);
-    
-    const [searchResults, setSearchResults] = useState(null);
-    const [translatedWord, setTranslatedWord] = useState("");
-    const [isLoading, setIsLoading] = useState(false);
-
-
     /*
     // 검색 버튼 클릭 시, 
     useEffect(() => {
@@ -33,37 +22,13 @@ export default function MainPage() {
     }, [isSearch]);
     */
 
-    const handleSearch = async () => {
-        if (!searchTerm.trim()) return; // 빈 값이면 검색 안 함
-        setIsLoading(true); // 로딩 시작
-
-        const { results, translatedText } = await getSearch(searchTerm);
-        setSearchResults(results || []);
-        setTranslatedWord(translatedText || "");
-
-        setIsLoading(false); // 로딩 끝
-    };
-
 
     return(
         <div className={styled['mainPage--container']}>
             <SearchContainer 
                 searchTerm={searchTerm} setSearchTerm={setSearchTerm}
-                onSearch={handleSearch} 
             />
-            {isLoading ? ( 
-                <BlankNews message="검색 중입니다..." />  
-            ) : searchResults === null ? ( 
-                <MainNews /> 
-            ) : searchResults.length > 0 ? (
-                <SearchNews 
-                    searchTerm={searchTerm} 
-                    searchResults={searchResults} 
-                    translatedText={translatedWord}
-                />
-            ) : (
-                <BlankNews message="검색 결과가 없습니다." />
-            )}                 
+            <MainNews />             
         </div>
     )
 }
